Renaming attributes of the planets in the data file
In the updated version of the exoplanet data, the short forms of the attributes are different. The planetary_system_io.py works for the attribute names shown in the previous data file. So I had to change the short forms of the attributes of the updated file to that of the previous file